+++ backend/src/config/prisma.ts
@@ -1,8 +1,9 @@
 import { PrismaClient } from "../prisma/generated/prisma/client.js";
 import { PrismaPg } from "@prisma/adapter-pg";
+import { envConfig } from "./env.js";
 
 const adapter = new PrismaPg({
-  connectionString: process.env.DATABASE_URL
+  connectionString: envConfig.databaseUrl
 });
 
 export const prisma = new PrismaClient({ adapter });