+++ backend/src/prisma/migrations/20260417184627_remove_account_sessions/migration.sql
@@ -1,0 +1,3 @@
+-- This drops the account sessions table since session management to moving to redis (>= v2.2.6)
+-- DropTable
+DROP TABLE "account_sessions";